Created a new common package, suited for placing shared level editor class components there. Decided to tidy up the editor a bit.
