+++ common/src/main/java/editor/EditorFileChooser.java
@@ -1,7 +1,7 @@
 /**
  * 
  */
-package common;
+package editor;
 
 import java.io.File;
